@@ -6,11 +6,11 @@
     "license": "na",
     "require": {
         "php": ">=7.1",
-        "laravel/framework": "5.4.*",
+        "laravel/framework": "5.5.*",
         "ryanwinchester/hubspot-php": "~1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5.7"
+        "phpunit/phpunit": "~6.0"
     },
     "autoload": {
         "psr-4": {
